+++ system/lib/libc/musl/src/select/select.c
@@ -82,7 +82,7 @@ static int emscripten_select(int nfds, fd_set *readfds, fd_set *writefds, fd_set
 		}
 	}
 
-	int rtn = __syscall_poll((intptr_t)fds, n, timeout);
+	int rtn = __syscall_poll(fds, n, timeout);
 	if (rtn < 0) {
 		free(fds);
 		return -1;